Moved router the mod.rs for user and group
